requirements.txt
Flask==3.0.0
plivo==4.45.0
python-dotenv==1.0.1
